requirements.txt
python-telegram-bot==13.15
sqlalchemy
flask
reportlab
python-dotenv
psycopg2-binary
urllib3<2.0
six
setuptools
fpdf2
